dashboard/requirements.txt
dash
pandas==2.1.0
python-dotenv == 0.21.1
dash-auth==2.0.0
dash_bootstrap_components==1.5.0